Add aimec plots to report and move atomic funcs to aimectools
