+++ avaframe/data/avaDamBreak/Inputs/damBreak_com1DFACfg.ini
@@ -2,26 +2,30 @@
 ## Copy to  local_com1DFACfg.ini and set you parameters
 # This file is part of Avaframe.
 
-[DEVFLAGS]
-debugPlot = False
 
 [GENERAL]
 # acceleration of gravity [m/s²]
 gravAcc = 9.81
 # density of snow [kg/m³]
 rho = 200
-rhoEnt = 200
+rhoEnt = 100
 # mass per particle [kg]
 massPerPart = 150
 # minimum flow depth [m]
 hmin = 0.05
+
 #+++++++++++++SPH parameters
 # SPH gradient option
+# 1) SamosAT style (no reprojecion on the surface, dz = 0 and gz is used)
+# 2) SamostAt but done in the local coord system (will hopefully allow us to add the earth pressure coef)
+# and this time reprojecion on the surface, dz not 0 and g3 is used
+# 3) same as 2) but correcting the gradient to get constant exact gradients
+# 4) same as 2) but correcting the gradient to get linear exact gradients
 sphOption = 1
 # minimum SPH distance [m]
 minRKern = 0.001
 
-velMagMin = 1.0e-6
+velMagMin = 1.0e-3
 depMin = 1.0e-6
 
 # interpolation option
@@ -30,11 +34,31 @@ depMin = 1.0e-6
 #                       -2: bilinear interpolation
 interpOption = 2
 
+# Normal computation on rectangular grid
+# 4 triangles method        6 triangles method         8 triangles method
+# +----U----UR---+---+--... +----+----+----+---+--... +----+----+----+---+--...
+# |   /|\   |   /|          |   /| 2 /|   /|          |\ 2 | 3 /|   /|           Y
+# |  / | \  |  / |          |  / |  / |  / |          | \  |  / |  / |           ^
+# | /  |  \ | /  | /        | /  | /  | /  | /        |  \ | /  | /  | /         |
+# |/ 1 | 2 \|/   |/         |/ 1 |/ 3 |/   |/         | 1 \|/ 4 |/   |/          |
+# +----P----L----+---+--... +----*----+----+---+--... +----*----+----+----+--... +-----> X
+# |\ 4 | 3 /|   /|          | 6 /| 4 /|   /|          | 8 /|\ 5 |   /|
+# | \  |  / |  / |          |  / |  / |  / |          |  / | \  |  / |
+# |  \ | /  | /  | /        | /  | /  | /  | /        | /  |  \ | /  | /
+# |   \|/   |/   |/         |/ 5 |/   |/   |/         |/ 7 | 6 \|/   |/
+# +----+----+----+---+--... +----+----+----+---+--... +----+----+----+---+--...
+# |   /|   /|   /|          |   /|   /|   /|          |   /|   /|   /|
+# 4 Options available : -1: simple cross product method (with the diagonals P-UR and U-L)
+#                       -4: 4 triangles method
+#                       -6: 6 triangles method
+#                       -8: 8 triangles method
+methodMeshNormal = 1
+
 # subgridMixingFactor
 subgridMixingFactor = 100.
 
 # release depth
-relTh = 5.0
+relTh = 1.
 
 #++++++++++++Time stepping parameters
 # courant number
@@ -48,7 +72,7 @@ dt = 0.01
 # save every [s]
 dtSave = 5
 # End time [s]
-Tend = 6
+tEnd = 6
 # stopCriterion (0.01, stops when ke<0.01*pke)
 stopCrit = 0.01
 
@@ -63,11 +87,15 @@ R = 0.05
 B = 4.13
 
 #++++++++++++ Resistance force parameters
+# height of the obstacles
 hRes = 10
+# resistance coeff
 cw = 0.5
+# diameter of obstacles
 dRes = 0.3
+# spacing between obstacles
 sres = 5
-# entrainment Erosion Energy
+#++++++++++++ entrainment Erosion Energy
 hEnt = 0.3
 entEroEnergy = 5000
 entShearResistance = 0
@@ -80,12 +108,18 @@ resType = FD_FV
 tSteps = 0|5
 
 
+#++++++++++++++++ particle Initialisation
+# if true use file (import particles initial distribution from file)
+initialiseParticlesFromFile = False
+particleFile =
 
 [FLAGS]
-# True if entrainment areas shall be included
-flagEnt = True
-# True if resistance areas shall be included
-flagRes = True
+# True if entrainment areas shall NOT be included
+noEntrainment = False
+# True if resistance areas shall NOT be included
+noResistance = False
+# True if only the entrainment resistance simulation should be done
+onlyEntrRes = False
 # True if dev release areas
 flagDev = False
 # specify a particular release area scenario, provide name of shapefile with or without extension .shp (optional)
@@ -99,6 +133,7 @@ plotFields = ppr_pfd_pfv
 unitppr = kPa
 unitpfd = m
 unitpfv = ms-1
+
 
 
 [DAMBREAK]
