+++ benchmarks/avaFlatPlaneNullTest/avaFlatPlane_AIMECPyCfg.ini
@@ -27,7 +27,7 @@ interpMethod = bilinear
 distance = 10
 
 # computational module for comparison separated by |
-comModules = ref|com1DFAPy
+comModules = benchmarkReference|com1DFAPy
 
 #---------------------------------------
