+++ benchmarks/avaHelixChannelEntTest/avaHelixChannel_AIMECCfg.ini
@@ -1,9 +1,6 @@
 ### Config File - This file contains the main settings for the ana3AIMEC run
 ## Set your parameters
 # This file is part of Avaframe.
-# This file will be overridden by local_ana3AIMEC.ini if it exists
-# So copy this file to local_ana3AIMEC.ini, adjust your variables there
-
 
 # Required settings------------------------------
 [INPATH]
@@ -29,7 +26,7 @@ interpMethod = bilinear
 distance = 10
 
 # computational module for comparison separated by |
-comModules = ref|com1DFA
+comModules = benchmarkReference|com1DFA
 
 #---------------------------------------
