+++ benchmarks/avaHitNullTest/avaHit_AIMECCfg.ini
@@ -1,9 +1,6 @@
 ### Config File - This file contains the main settings for the ana3AIMEC run
 ## Set your parameters
 # This file is part of Avaframe.
-# This file will be overridden by local_ana3AIMEC.ini if it exists
-# So copy this file to local_ana3AIMEC.ini, adjust your variables there
-
 
 # Required settings------------------------------
 [INPATH]
@@ -14,7 +11,7 @@
 # Optional settings-------------------------------
 [AIMECSETUP]
 # width of the domain around the avalanche path in [m]
-domainWidth = 600
+domainWidth = 1200
 
 # angle for the start of the run-out zone
 startOfRunoutAngle = 10
@@ -29,7 +26,7 @@ interpMethod = bilinear
 distance = 10
 
 # computational module for comparison separated by |
-comModules = ref|com1DFA
+comModules = benchmarkReference|com1DFA
 
 #---------------------------------------
