+++ benchmarks/avaHitNullTest/avaHit_AIMECPyCfg.ini
@@ -12,7 +12,7 @@
 # Optional settings-------------------------------
 [AIMECSETUP]
 # width of the domain around the avalanche path in [m]
-domainWidth = 600
+domainWidth = 1200
 
 # angle for the start of the run-out zone
 startOfRunoutAngle = 10
@@ -27,7 +27,7 @@ interpMethod = bilinear
 distance = 10
 
 # computational module for comparison separated by |
-comModules = ref|com1DFAPy
+comModules = benchmarkReference|com1DFAPy
 
 #---------------------------------------
